Trait hotshot_task::event_stream::EventStream

pub trait EventStream: Clone + 'static + Sync + Send {
    type EventType: PassType;
    type StreamType: Stream<Item = Self::EventType> + Sync + Send + 'static;

    // Required methods
    fn publish<'life0, 'async_trait>(
        &'life0 self,
        event: Self::EventType
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 self,
        filter: FilterEvent<Self::EventType>
    ) -> Pin<Box<dyn Future<Output = (Self::StreamType, StreamId)> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unsubscribe<'life0, 'async_trait>(
        &'life0 self,
        id: StreamId
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Async pub sub event stream NOTE: static bound indicates that if the type points to data, that data lives for the lifetime of the program

Required Associated Types

type EventType: PassType

the type of event to process

type StreamType: Stream<Item = Self::EventType> + Sync + Send + 'static

the type of stream to use

Required Methods

fn publish<'life0, 'async_trait>( &'life0 self, event: Self::EventType ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

publish an event to the event stream

fn subscribe<'life0, 'async_trait>( &'life0 self, filter: FilterEvent<Self::EventType> ) -> Pin<Box<dyn Future<Output = (Self::StreamType, StreamId)> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

subscribe to a particular set of events specified by filter. Filter returns true if the event should be propagated

fn unsubscribe<'life0, 'async_trait>( &'life0 self, id: StreamId ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

unsubscribe from the stream