Struct hotshot_task::task::HST

pub struct HST<HSTT: HotShotTaskTypes> {

    pub(crate) stream_id: Option<StreamId>,
    r_val: Option<HotShotTaskCompleted>,
    in_progress_shutdown_fut: Option<BoxFuture<'static, ()>>,
    in_progress_fut: Option<BoxFuture<'static, (Option<HotShotTaskCompleted>, HSTT::State)>>,
    name: String,
    status: TaskState,
    shutdown_fns: Vec<ShutdownFn>,
    event_stream: MaybePinnedEventStream<HSTT>,
    message_stream: Option<Pin<Box<Fuse<HSTT::MessageStream>>>>,
    state: Option<HSTT::State>,
    handle_event: Option<HandleEvent<HSTT>>,
    handle_message: Option<HandleMessage<HSTT>>,
    pub(crate) tid: Option<HotShotTaskId>,

}

hot shot task

Fields
stream_id: Option<StreamId>

Optional ID of the stream.

r_val: Option<HotShotTaskCompleted>

the eventual return value, post-cleanup

in_progress_shutdown_fut: Option<BoxFuture<'static, ()>>

if we have a future for tracking shutdown progress

in_progress_fut: Option<BoxFuture<'static, (Option<HotShotTaskCompleted>, HSTT::State)>>

the in progress future

name: String

name of task

status: TaskState

state of the task TODO make this boxed. We don’t want to assume this is a small future. since it concievably may be stored on the stack

shutdown_fns: Vec<ShutdownFn>

functions performing cleanup one should shut down the task if we’re tracking with a global registry the other should unsubscribe from the stream

event_stream: MaybePinnedEventStream<HSTT>

shared stream

message_stream: Option<Pin<Box<Fuse<HSTT::MessageStream>>>>

stream of messages

state: Option<HSTT::State>

state

handle_event: Option<HandleEvent<HSTT>>

handler for events

handle_message: Option<HandleMessage<HSTT>>

handler for messages

tid: Option<HotShotTaskId>

task id

Implementations

impl<HSTT: HotShotTaskTypes> HST<HSTT>

pub(crate) fn project<'pin>( self: Pin<&'pin mut Self> ) -> ProjectedHST<'pin, HSTT>

pub(crate) fn project_ref<'pin>( self: Pin<&'pin Self> ) -> __HSTProjectionRef<'pin, HSTT>

impl<HSTT: HotShotTaskTypes> HST<HSTT>

pub(crate) fn base_check(&self)

Do a consistency check on the HST construction

pub(crate) fn event_check(&self)

perform event sanity checks

pub(crate) fn message_check(&self)

perform message sanity checks

pub(crate) fn register_handler(self, handler: HotShotTaskHandler<HSTT>) -> Self

register a handler with the task

pub(crate) async fn register_event_stream( self, event_stream: HSTT::EventStream, filter: FilterEvent<HSTT::Event> ) -> Self

register an event stream with the task

pub(crate) fn register_message_stream(self, stream: HSTT::MessageStream) -> Self

register a message with the task

pub(crate) fn register_state(self, state: HSTT::State) -> Self

register state with the task

pub(crate) async fn register_registry( self, registry: &mut GlobalRegistry ) -> Self

register with the registry

pub(crate) fn new(name: String) -> Self

create a new task

pub fn launch(self) -> BoxFuture<'static, HotShotTaskCompleted>

launch the task NOTE: the only way to get a HST is by usage of one of the impls. Those all have checks enabled. So, it should be safe to launch.