Struct hotshot_task::task::Task

pub struct Task<S: TaskState> {
    event_sender: Sender<S::Event>,
    event_receiver: Receiver<S::Event>,
    registry: Arc<TaskRegistry>,
    state: S,
}

A basic task which loops waiting for events to come from event_receiver and then handles them using it’s state It sends events to other Tasks through event_sender This should be used as the primary building block for long running or medium running tasks (i.e. anything that can’t be described as a dependency task)

Fields
event_sender: Sender<S::Event>

Sends events all tasks including itself

event_receiver: Receiver<S::Event>

Receives events that are broadcast from any task, including itself

registry: Arc<TaskRegistry>

Contains this task, used to register any spawned tasks

state: S

The state of the task. It is fed events from event_sender and mutates it state ocordingly. Also it signals the task if it is complete/should shutdown

Implementations

impl<S: TaskState + Send + 'static> Task<S>

pub fn new( tx: Sender<S::Event>, rx: Receiver<S::Event>, registry: Arc<TaskRegistry>, state: S ) -> Self

Create a new task

pub fn handle_event( &mut self, event: S::Event ) -> impl Future<Output = Option<S::Output>> + Send + '_
where Self: Sized,

The Task analog of TaskState::handle_event.

pub fn run(self) -> JoinHandle<()>

Spawn the task loop, consuming self. Will continue until the task reaches some shutdown condition

pub fn subscribe(&self) -> Receiver<S::Event>

Create a new event Receiver from this Task’s receiver. The returned receiver will get all messages not yet seen by this task

pub fn sender(&self) -> &Sender<S::Event>

Get a new sender handle for events

pub fn clone_sender(&self) -> Sender<S::Event>

Clone the sender handle

pub async fn send( &self, event: S::Event ) -> Result<Option<S::Event>, SendError<S::Event>>

Broadcast a message to all listening tasks

Errors

Errors if the broadcast fails

pub fn state_mut(&mut self) -> &mut S

Get a mutable reference to this tasks state

pub fn state(&self) -> &S

Get an immutable reference to this tasks state

pub async fn run_sub_task(&self, state: S)

Spawn a new task adn register it. It will get all events not seend by the task creating it.