Trait hotshot_task::task::HotShotTaskTypes

pub trait HotShotTaskTypes: 'static {
    type Event: PassType;
    type State: TS;
    type EventStream: EventStream<EventType = Self::Event>;
    type Message: PassType;
    type MessageStream: SendableStream<Item = Self::Message>;
    type Error: TaskErr + 'static + ?Sized;

    // Required method
    fn build(builder: TaskBuilder<Self>) -> HST<Self> 
       where Self: Sized;
}

group of types needed for a hotshot task

Required Associated Types

type Event: PassType

the event type from the event stream

type State: TS

the state of the task

type EventStream: EventStream<EventType = Self::Event>

the global event stream

type Message: PassType

the message stream to receive

type MessageStream: SendableStream<Item = Self::Message>

the steam of messages from other tasks

type Error: TaskErr + 'static + ?Sized

the error to return

Required Methods

fn build(builder: TaskBuilder<Self>) -> HST<Self> where Self: Sized,

build a task NOTE: done here and not on TaskBuilder because we want specific checks done on each variant NOTE: all generics implement Sized, but this bound is NOT applied to Self unless we specify