Trait hotshot_task::task::TaskEvent

pub trait TaskEvent: PartialEq {
    // Required method
    fn shutdown_event() -> Self;
}

Trait for events that long-running tasks handle

Required Methods

fn shutdown_event() -> Self

The shutdown signal for this event type

Note that this is necessarily uniform across all tasks. Exiting the task loop is handled by the task spawner, rather than the task individually.

Object Safety

This trait is not object safe.