Trait hotshot_task::task::TestTaskState

pub trait TestTaskState: Send {
    type Message: Clone + Send + Sync + 'static;
    type Output: Send;
    type State: TaskState;

    // Required method
    fn handle_message(
        message: Self::Message,
        id: usize,
        task: &mut TestTask<Self::State, Self>
    ) -> impl Future<Output = Option<Self::Output>> + Send
       where Self: Sized;
}

Task state for a test. Similar to TaskState but it handles messages as well as events. Messages are events that are external to this task. (i.e. a test message would be an event from non test task) This is used as state for TestTask and messages can come from many different input streams.

Required Associated Types

type Message: Clone + Send + Sync + 'static

Message type handled by the task

type Output: Send

Result returned by the test task on completion

type State: TaskState

The state type

Required Methods

fn handle_message( message: Self::Message, id: usize, task: &mut TestTask<Self::State, Self> ) -> impl Future<Output = Option<Self::Output>> + Send
where Self: Sized,

Handle and incoming message and return Some if the task is finished