Trait hotshot_task::dependency::Dependency

pub trait Dependency<T> {
    // Required method
    fn completed(self) -> impl Future<Output = Option<T>> + Send;

    // Provided methods
    fn or<D: Dependency<T> + Send + 'static>(self, dep: D) -> OrDependency<T>
       where T: Send + Sync + Clone + 'static,
             Self: Sized + Send + 'static { ... }
    fn and<D: Dependency<T> + Send + 'static>(self, dep: D) -> AndDependency<T>
       where T: Send + Sync + Clone + 'static,
             Self: Sized + Send + 'static { ... }
}

Type which describes the idea of waiting for a dependency to complete

Required Methods

fn completed(self) -> impl Future<Output = Option<T>> + Send

Complete will wait until it gets some value T then return the value

Provided Methods

fn or<D: Dependency<T> + Send + 'static>(self, dep: D) -> OrDependency<T>
where T: Send + Sync + Clone + 'static, Self: Sized + Send + 'static,

Create an or dependency from this dependency and another

fn and<D: Dependency<T> + Send + 'static>(self, dep: D) -> AndDependency<T>
where T: Send + Sync + Clone + 'static, Self: Sized + Send + 'static,

Create an and dependency from this dependency and another

Object Safety

This trait is not object safe.